outsourced TextRenderer
