+++ app/lib/objectivemodeler/draw/index.js
@@ -1,5 +1,5 @@
 import ODRenderer from './ODRenderer';
-import TextRenderer from './TextRenderer';
+import TextRenderer from '../../frequentlyUsedFiles/draw/TextRenderer.js';
 
 export default {
   __init__: [ 'odRenderer' ],
