+++ app/lib/resourcemodeler/draw/index.js
@@ -1,5 +1,5 @@
 import REMRenderer from './REMRenderer';
-import TextRenderer from './TextRenderer';
+import TextRenderer from '../../frequentlyUsedFiles/draw/TextRenderer.js';
 
 export default {
   __init__: [ 'remRenderer' ],
